apvs-340 - moved unit tests to unit folder
