+++ test/unit/routes/mock-view-engine.js
@@ -6,7 +6,7 @@ var viewEngine = function (app, viewsPath) {
     return callback(null, rendered)
   })
   app.set('view engine', 'html')
-  app.set('views', [ path.join(__dirname, viewsPath), path.join(__dirname, '../../lib/') ])
+  app.set('views', [ path.join(__dirname, viewsPath), path.join(__dirname, '../../../lib/') ])
 }
 
 exports.default = viewEngine
