+++ test/unit/routes/test-route-status.js
@@ -1,7 +1,7 @@
 /* global describe beforeEach it */
 var supertest = require('supertest')
 var express = require('express')
-var route = require('../../app/routes/status')
+var route = require('../../../app/routes/status')
 
 describe('status', function () {
   var request
